+++ XFaction/Core/Locales/enUS.lua
@@ -114,6 +114,7 @@ L['CHAT_ONLINE_SOUND'] = 'Play Sound'
 L['CHAT_ONLINE_SOUND_TOOLTIP'] = 'Play sound when any confederate member comes online'
 L['CHAT_LOGIN'] = 'has come online.'
 L['CHAT_LOGOUT'] = 'has gone offline.'
+L['CHAT_NO_PLAYER_FOUND'] = 'No player named '
 --=========================================================================
 -- DataText Specific
 --=========================================================================